* fixed eating a little

+++ antargis/trunk/src/entity.cc
@@ -330,32 +330,32 @@ void AntEntity::setDirection(float dir)
 Resource::Resource()
 {
 }
-int Resource::get(const std::string &pName)
+float Resource::get(const std::string &pName)
 {
   return r[pName];
 }
-void Resource::add(const std::string &pName,int value)
+void Resource::add(const std::string &pName,float value)
 {
   r[pName]+=value;
 }
-void Resource::sub(const std::string &pName,int value)
+void Resource::sub(const std::string &pName,float value)
 {
-  r[pName]=std::max(r[pName]-value,0);
+  r[pName]=std::max(r[pName]-value,0.0f);
 }
-void Resource::set(const std::string &pName,int value)
+void Resource::set(const std::string &pName,float value)
 {
   r[pName]=value;
 }
 
 void Resource::takeAll(Resource &pr)
 {
-  std::map<std::string,int>::iterator i=pr.r.begin();
+  std::map<std::string,float>::iterator i=pr.r.begin();
   for(;i!=pr.r.end();i++)
     add(i->first,i->second);
   pr.r.clear();
 }
 
-std::map<std::string,int> Resource::getAll() const
+std::map<std::string,float> Resource::getAll() const
 {
   return r;
 }
